fix: use i32 for SELECT 1 readiness check
PostgreSQL returns integer (4 bytes) for the literal 1; sqlx decode into
i64 (bigint) fails at runtime. Fixes /ready returning "not ready" when
the DB is reachable.

+++ services/api-gateway/src/routes/health.rs
@@ -10,7 +10,7 @@ pub fn router() -> Router<AppState> {
 
 async fn ready(axum::extract::State(state): axum::extract::State<AppState>) -> &'static str {
     // Cheap readiness check: ping the DB.
-    match sqlx::query_scalar::<_, i64>("SELECT 1")
+    match sqlx::query_scalar::<_, i32>("SELECT 1")
         .fetch_one(&state.db)
         .await
     {
